TJShow: - Installer is now multi-lingual and writes 'initial setting files' to the user's profile which select the correct locale (en or nl) - Removed some unneeded resource bmp's (obsolete since almost all dialogs are gone now) - Fixed syntax error in demo-only code (such a shame...) - GetSizeString now uses a 64-bit number so large files are handled correctly - FileWnd has an extra 'size' column, and the icons are now loaded the modern way
TJDMX:
- Cleaned up a bit

TJShared:
- Fixed problems with right clicking in a ListWnd
- ToolbarWnd now hides stuff if it is too small


git-svn-id: https://svn.pixelspark.nl/svn/corespark/trunk@288 c9a305e2-9f8a-4be3-8c95-b0e5bab41dc9

+++ TJShared/include/tjfile.h
@@ -3,7 +3,7 @@
 
 namespace tj {
 	namespace shared {
-		typedef unsigned int Bytes;
+		typedef long long Bytes; // This is equivalent to __int64 on MSVC++
 
 		class EXPORTED File {
 			public:
